TP_2/requirements.txt
# Servidor Async y cliente HTTP
aiohttp

# Parsing
beautifulsoup4
lxml

# Procesamiento de imágenes
Pillow

# Procesamiento (Screenshot y Performance)
selenium
webdriver-manager
requests

# --- Testing ---
pytest
pytest-asyncio